Add occurences_without_exdates
Useful for when you want to exclude EXDATE occurences from an rruleset. IE, you cancel or modify a recurring event, and are generating dates in a view, you'd ideally want to add an EXDATE and have it no longer be present (replaced by the new event, in the case of a modification).